\name{packageNavPill}
\alias{packageNavPill}
\title{Create Navigation Links for Header for Package Docs}
\usage{
packageNavPill(githubHref, docsActive = TRUE)
}
\arguments{
  \item{githubHref}{URL pointing to github documentation}

  \item{docsActive}{should the docs link be active?}
}
\description{
Create navigation header links specifically for package
documentation (links include "Docs", "Function Ref", and
"Github")
}

